Encode as Alpha if any frames are non-opaque
If an APNG to be encoded is in an RGBA/NRGBA or similar format, we now
check all frames rather than just the first to see if they are opaque or not. This fixes issues wherein an APNG with a fully opaque first frame would encode partially opaque subsequent frames as opaque as well, rendering all transparency as opaque.